[Restructuring] no import shortcuts.
