+++ src/semantic_world/adapters/multi_parser.py
@@ -27,9 +27,9 @@ except ImportError as e:
     JointType = None
     UsdUrdf = None
 
-from ..world_description import RevoluteConnection, PrismaticConnection, FixedConnection
-from ..world_description import DegreeOfFreedom
-from ..datastructures import PrefixedName
+from ..world_description.connections import RevoluteConnection, PrismaticConnection, FixedConnection
+from ..world_description.degree_of_freedom import DegreeOfFreedom
+from ..datastructures.prefixed_name import PrefixedName
 from ..spatial_types import spatial_types as cas
 from ..spatial_types.derivatives import DerivativeMap
 from ..world import World, Body, Connection
