--- src/semantic_world/datastructures/__init__.py
@@ -1,3 +1,0 @@
-from .prefixed_name import PrefixedName
-from .types import AnyMatrix4x4, NpMatrix4x4
-from .variables import SpatialVariables
