+++ src/semantic_world/orm/model.py
@@ -8,14 +8,15 @@ import trimesh.exchange.stl
 from ormatic.dao import AlternativeMapping
 from sqlalchemy import TypeDecorator, types
 
-from ..world_description import DegreeOfFreedom
-from ..datastructures import PrefixedName
+from ..world_description.degree_of_freedom import DegreeOfFreedom
+from ..datastructures.prefixed_name import PrefixedName
 from ..spatial_types import RotationMatrix, Vector3, Point3, TransformationMatrix
 from ..spatial_types.derivatives import DerivativeMap
 from ..spatial_types.spatial_types import Quaternion
 from ..spatial_types.symbol_manager import symbol_manager
 from ..world import World, Body
-from ..world_description import Connection, View
+from ..world_description.connections import Connection
+from ..world_description.world_entity import View
 
 
 @dataclass
