+++ src/semantic_world/reasoning/world_rdr/world_views_mcrdr_defs.py
@@ -1,7 +1,7 @@
 from typing_extensions import List, Union
 
 from ...views.views import Cabinet, Container, Door, Drawer, Fridge, Handle
-from ...world_description import FixedConnection, PrismaticConnection, RevoluteConnection
+from ...world_description.connections import FixedConnection, PrismaticConnection, RevoluteConnection
 from ...world import World
