+++ src/semantic_world/world_description/__init__.py
@@ -1,13 +1,2 @@
-from .connections import (FixedConnection, PrismaticConnection, RevoluteConnection,
-                          Connection6DoF, ActiveConnection, PassiveConnection, OmniDrive)
-from .connection_factories import (ConnectionFactory, FixedConnectionFactory, PrismaticConnectionFactory,
-                                   RevoluteConnectionFactory)
-from .world_entity import Body, KinematicStructureEntity, Region, View, Connection, EnvironmentView
-from .geometry import (Mesh, TriangleMesh, BoundingBox, BoundingBoxCollection, Scale,
-                       Box, Sphere, Cylinder, Primitive, Shape)
-from .degree_of_freedom import DegreeOfFreedom
-from .world_modification import (WorldModelModificationBlock, WorldModelModification, UnknownWorldModification,
-                                 AddBodyModification, AddConnectionModification, AddDegreeOfFreedomModification)
-from .graph_of_convex_sets import GraphOfConvexSets, PoseOccupiedError
-from .world_state import WorldState
+
