+++ test/test_ros/test_world_modifications.py
@@ -1,10 +1,10 @@
 import unittest
 
-from semantic_world.world_description import (
+from semantic_world.world_description.connection_factories import (
     ConnectionFactory,
     FixedConnectionFactory,
 )
-from semantic_world.world_description import (
+from semantic_world.world_description.connections import (
     FixedConnection,
     Connection6DoF,
     PrismaticConnection,
@@ -12,8 +12,8 @@ from semantic_world.world_description import (
 from semantic_world.datastructures.prefixed_name import PrefixedName
 from semantic_world.spatial_types.spatial_types import Vector3
 from semantic_world.world import World
-from semantic_world.world_description import Body
-from semantic_world.world_description import (
+from semantic_world.world_description.world_entity import Body
+from semantic_world.world_description.world_modification import (
     WorldModelModificationBlock,
     UnknownWorldModification,
     AddBodyModification,
