+++ test/test_worlds/test_gcs.py
@@ -6,12 +6,12 @@ from random_events.interval import SimpleInterval
 from random_events.product_algebra import SimpleEvent
 
 from semantic_world.adapters.urdf import URDFParser
-from semantic_world.world_description import BoundingBox, BoundingBoxCollection
-from semantic_world.world_description import GraphOfConvexSets, PoseOccupiedError
+from semantic_world.world_description.geometry import BoundingBox, BoundingBoxCollection
+from semantic_world.world_description.graph_of_convex_sets import GraphOfConvexSets, PoseOccupiedError
 from semantic_world.spatial_types import Point3
 from semantic_world.datastructures.variables import SpatialVariables
 from semantic_world.world import World
-from semantic_world.world_description import Body
+from semantic_world.world_description.world_entity import Body
 
 
 class GCSTestCase(unittest.TestCase):
